move lock helpers to own file
